First run particles

+++ src/main/java/home/mutant/dl/particlefilter/Particle.java
@@ -1,7 +1,7 @@
 package home.mutant.dl.particlefilter;
 
 public class Particle {
-	float x;
-	float y;
-	float weight;
+	int x;
+	int y;
+	double weight;
 }
